pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "numpy>=2.3.1",
    "pandas>=2.3.1",
    "pathlib>=1.0.1",
    "python-dotenv>=1.1.1",
    "upsonic>=0.60.0",
    "yfinance>=0.2.65",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
]
type = [
    "mypy>=1.15.0",
]


[tool.ruff] 
line-length=80

[tool.mypy]
plugins = ["pydantic.mypy"]
ignore_missing_imports = true